+++ gen/com/intellij/idea/plugin/hybris/impex/psi/ImpexDocumentIdDec.java
@@ -4,7 +4,7 @@
  * ----------------------------------------------------------------
  *
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -24,7 +24,11 @@ package com.intellij.idea.plugin.hybris.impex.psi;
 import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
+import java.util.Collection;
+import java.util.Map;
 
 public interface ImpexDocumentIdDec extends ImpexPsiNamedElement {
+
+  @NotNull Map<String, Collection<ImpexValue>> getValues();
 
 }