+++ src/com/intellij/idea/plugin/hybris/impex/find/findUsages/ImpexFindUsagesProvider.kt
@@ -1,10 +1,11 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as 
- * published by the Free Software Foundation, either version 3 of the 
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
@@ -34,7 +35,8 @@ class ImpexFindUsagesProvider : FindUsagesProvider {
         TokenSet.orSet(
             TokenSet.create(ImpexTypes.MACRO_NAME_DECLARATION),
             TokenSet.create(ImpexTypes.MACRO_DECLARATION),
-            TokenSet.create(ImpexTypes.MACRO_USAGE)
+            TokenSet.create(ImpexTypes.MACRO_USAGE),
+            TokenSet.create(ImpexTypes.FIELD_VALUE),
         ),
         TokenSet.create(
             ImpexTypes.COMMENT,