+++ src/com/intellij/idea/plugin/hybris/system/type/meta/model/TSMetaEnum.kt
@@ -18,9 +18,11 @@
  */
 package com.intellij.idea.plugin.hybris.system.type.meta.model
 
+import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.lang.documentation.renderer.hybrisDoc
 import com.intellij.idea.plugin.hybris.system.type.model.EnumType
 import com.intellij.idea.plugin.hybris.system.type.model.EnumValue
+import javax.swing.Icon
 
 interface TSMetaEnum : TSMetaClassifier<EnumType> {
     val values: Map<String, TSMetaEnumValue>
@@ -38,6 +40,8 @@ interface TSMetaEnum : TSMetaClassifier<EnumType> {
 
 interface TSGlobalMetaEnum : TSMetaEnum, TSGlobalMetaClassifier<EnumType>, TSTypedClassifier {
     override val declarations: MutableSet<TSMetaEnum>
+    override val icon: Icon
+        get() = HybrisIcons.TypeSystem.Types.ENUM
 
     override fun documentation() = hybrisDoc {
         title("Enum type", name ?: "?")